Don't add the border when heightOffset is not present
Earlier we were always adding border of 1px on either side when frame
was not present. But for tripgain, an responsive css breaks the layout
as the size of captured iframe (during recording) is different than size
of rendered iframe; even if it was for 2px. Hence temporarily we removed
the border.